Conflicts:
	Trakttv.bundle/Contents/Code/core/plugin.py

+++ Trakttv.bundle/Contents/Code/pts/scrobbler_websocket.py
@@ -12,7 +12,6 @@ log = Logger('pts.scrobbler_websocket')
 
 class WebSocketScrobbler(ScrobblerMethod):
     name = 'WebSocketScrobbler'
-    legacy = False
 
     def __init__(self):
         super(WebSocketScrobbler, self).__init__()
@@ -47,19 +46,34 @@ class WebSocketScrobbler(ScrobblerMethod):
 
         log.debug('Creating a WatchSession for the current media')
 
-        video_section = PlexMediaServer.get_session(session_key)
-        if not video_section:
+        skip = False
+
+        info = PlexMediaServer.get_session(session_key)
+        if not info:
             return None
 
-        player_section = video_section.findall('Player')
+        # Client
+        player_section = info.findall('Player')
         if len(player_section):
             player_section = player_section[0]
 
-        session = WatchSession.from_section(
-            video_section, state,
-            PlexMetadata.get(video_section.get('ratingKey')).to_dict(),
-            PlexMediaServer.get_client(player_section.get('machineIdentifier'))
-        )
+        client = PlexMediaServer.get_client(player_section.get('machineIdentifier'))
+
+        # Metadata
+        metadata = None
+
+        try:
+            metadata = PlexMetadata.get(info.get('ratingKey'))
+
+            if metadata:
+                metadata = metadata.to_dict()
+        except NotImplementedError, e:
+            # metadata not supported (music, etc..)
+            log.debug('%s, ignoring session' % e.message)
+            skip = True
+
+        session = WatchSession.from_section(info, state, metadata, client)
+        session.skip = skip
         session.save()
 
         return session
@@ -75,15 +89,7 @@ class WebSocketScrobbler(ScrobblerMethod):
         log.debug('last item key: %s, current item key: %s' % (session.item_key, video_section.get('ratingKey')))
 
         if session.item_key != video_section.get('ratingKey'):
-            log.info('Invalid Session: Media changed')
-            return False
-
-        if not session.metadata:
-            log.debug('Invalid Session: Missing metadata')
-            return False
-
-        if session.metadata.get('duration', 0) <= 0:
-            log.debug('Invalid Session: Invalid duration')
+            log.debug('Invalid Session: Media changed')
             return False
 
         session.last_view_offset = view_offset
@@ -91,36 +97,78 @@ class WebSocketScrobbler(ScrobblerMethod):
 
         return True
 
+    def session_valid(self, session):
+        if not session.metadata:
+            if session.skip:
+                return True
+
+            log.debug('Invalid Session: Missing metadata')
+            return False
+
+        if session.metadata.get('duration', 0) <= 0:
+            log.debug('Invalid Session: Invalid duration')
+            return False
+
+        return True
+
     def get_session(self, session_key, state, view_offset):
         session = WatchSession.load(session_key)
 
-        if session:
-            if session.last_view_offset and session.last_view_offset > view_offset:
-                log.debug('View offset has gone backwards (last: %s, cur: %s)' % (
-                    session.last_view_offset, view_offset
-                ))
-
-                # First try update the session if the media hasn't changed
-                # otherwise delete the session
-                if not self.update_session(session, view_offset):
-                    log.debug('Media changed, deleting the session')
-                    session.delete()
-                    return None
-
-            if session.skip:
-                return None
-
-            if state == 'playing' and session.update_required:
-                log.debug('Session update required, updating the session...')
-
-                if not self.update_session(session, view_offset):
-                    log.debug('Media changed, deleting the session')
-                    session.delete()
-                    return None
-        else:
+        if not session:
             session = self.create_session(session_key, state)
 
+            if not session:
+                return None
+
+        update_session = False
+
+        # Update session when view offset goes backwards
+        if session.last_view_offset and session.last_view_offset > view_offset:
+            log.debug('View offset has gone backwards (last: %s, cur: %s)' % (
+                session.last_view_offset, view_offset
+            ))
+
+            update_session = True
+
+        # Update session on missing metadata + session skip
+        if not session.metadata and session.skip:
+            update_session = True
+
+        # First try update the session if the media hasn't changed
+        # otherwise delete the session
+        if update_session and not self.update_session(session, view_offset):
+            log.debug('Media changed, deleting the session')
+            session.delete()
+            return None
+
+        # Delete session if invalid
+        if not self.session_valid(session):
+            session.delete()
+            return None
+
+        if session.skip:
+            return None
+
+        if state == 'playing' and session.update_required:
+            log.debug('Session update required, updating the session...')
+
+            if not self.update_session(session, view_offset):
+                log.debug('Media changed, deleting the session')
+                session.delete()
+                return None
+
         return session
+
+    def valid(self, session):
+        # Check filters
+        if not self.valid_user(session) or\
+           not self.valid_client(session) or \
+           not self.valid_section(session):
+            session.skip = True
+            session.save()
+            return False
+
+        return True
 
     def update(self, session_key, state, view_offset):
         # Ignore if scrobbling is disabled
@@ -129,33 +177,15 @@ class WebSocketScrobbler(ScrobblerMethod):
 
         session = self.get_session(session_key, state, view_offset)
         if not session:
-            log.info('Invalid session, unable to continue')
+            log.trace('Invalid or ignored session, nothing to do')
             return
 
         # Ignore sessions flagged as 'skip'
         if session.skip:
             return
 
-        # Ensure we are only scrobbling for the myPlex user listed in preferences
-        if not self.valid_user(session):
-            log.info('Ignoring item [%s](%s) played by other user: %s' % (
-                session_key,
-                session.get_title(),
-                session.user.title if session.user else None
-            ))
-            session.skip = True
-
-        # Ensure we are only scrobbling for the client listed in preferences
-        if not self.valid_client(session):
-            log.info('Ignoring item [%s](%s) played by other client: %s' % (
-                session_key,
-                session.get_title(),
-                session.client.name if session.client else None
-            ))
-            session.skip = True
-
-        if session.skip:
-            session.save()
+        # Validate session (check filters)
+        if not self.valid(session):
             return
 
         media_type = session.get_type()
@@ -180,10 +210,7 @@ class WebSocketScrobbler(ScrobblerMethod):
         if action:
             self.handle_action(session, media_type, action, state)
         else:
-            log.debug('%s Nothing to do this time for %s' % (
-                self.get_status_label(session.progress, state),
-                session.get_title()
-            ))
+            log.debug(self.status_message(session, state)('Nothing to do this time for %s'))
             session.save()
 
         if self.handle_state(session, state) or action:
